feat: integrate JLVM gas metering into fiber orchestration

+++ modules/models/src/main/scala/xyz/kd5ujc/schema/Records.scala
@@ -51,7 +51,7 @@ object Records {
       reason:             String,
       attemptedAt:        SnapshotOrdinal,
       attemptedEventType: StateMachine.EventType,
-      gasUsed:            Int,
+      gasUsed:            Long,
       depth:              Int
     ) extends EventProcessingStatus
